feat: refactor sequence components and add text dragging functionality
- Separate sequence item components to fix hook order issues
- Add mouse drag functionality for text elements
- Create individual components for video, image, text, and audio items
- Implement text position updates on drag
- Improve code modularity and user interaction
